build on mac;try cross compile from linux using 'GOOS=darwin GOARCH=amd64 go build'
