+++ lepton/ldd_darwin.go
@@ -1,0 +1,7 @@
+package lepton
+
+func getSharedLibs(path string) ([]string, error) {
+	// DYLD_PRINT_LIBRARIES ? or parse
+	var deps []string
+	return deps, nil
+}
