+++ network_darwin.go
@@ -1,0 +1,15 @@
+package main
+
+import "github.com/vishvananda/netlink"
+
+func addTapDevice(name string, bridge *netlink.Bridge) error {
+	return nil
+}
+
+func dhcpAcquireIPAddress(mac string) (string, error) {
+	return "", nil
+}
+
+func setupBridgeNetwork() error {
+	return nil
+}
